made review changes
